+++ CHANGELOG.md
@@ -6,3 +6,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
+
+* Added Angular project `ts/frontend`
+* Added Node.js project `ts/backend`